+++ example/lib/example/models/get_example_detail_model.dart
@@ -1,3 +1,0 @@
-part of '../example.dart';
-
-class GetExampleDetailModel {}