+++ example/lib/example/models/send_example_update_model.dart
@@ -1,5 +1,0 @@
-part of '../example.dart';
-
-class SendExampleUpdateSuccessModel {}
-
-class SendExampleUpdateFailedModel {}