+++ example/lib/submit/post_controller.dart
@@ -5,7 +5,7 @@ class CreateUserController extends SendController {
   bool get autoDispose => false;
 
   @override
-  Future<void> request(Map<String, dynamic> args) async {
+  Future<void> onRequest(Map<String, dynamic> args) async {
     await Future.delayed(const Duration(seconds: 1));
     final response = await Dio().post(
       'https://reqres.in/api/users/2',
@@ -18,7 +18,7 @@ class CreateUserController extends SendController {
     );
 
     if (response.statusCode == 201) {
-      final model = CreateUserModel.fromJson(response.data);
+      final model = CreateUserModel.fromJSON(response.data);
       emit(SendSuccessState<CreateUserModel>(data: model));
     } else {
       emit(SendFailedState<Map<String, dynamic>>(