+++ lib/api_bloc.dart
@@ -3,6 +3,7 @@
 /// calls and manage state changes.
 library api_bloc;
 
+import 'package:equatable/equatable.dart';
 import 'package:flutter/material.dart';
 part 'package:api_bloc/src/controllers/bloc_controller.dart';
 part 'package:api_bloc/src/controllers/get_controller.dart';
@@ -10,4 +11,4 @@ part 'package:api_bloc/src/controllers/send_controller.dart';
 part 'package:api_bloc/src/states/get_states.dart';
 part 'package:api_bloc/src/states/send_states.dart';
 part 'package:api_bloc/src/states/bloc_states.dart';
-part 'package:api_bloc/src/widgets/bloc_widget.dart';
+part 'package:api_bloc/src/views/bloc_widget.dart';