+++ lib/src/controllers/get_controller.dart
@@ -5,9 +5,8 @@ part of 'package:api_bloc/api_bloc.dart';
 ///
 /// ```dart
 /// class GetUserController extends GetController {
-///
 ///   @override
-///   Future<void> request() async {
+///   Future<void> onRequest() async {
 ///     Response response = await Response.get('https://base.url/api/user',
 ///       onProgress: (double progress) {
 ///         emit(GetLoadingState<double>(data: progress));
@@ -27,7 +26,7 @@ part of 'package:api_bloc/api_bloc.dart';
 /// type to emphasize the data that we're going to use in [ApiBloc].
 /// ```dart
 /// // in controller
-/// Future<void> request() async {
+/// Future<void> onRequest() async {
 ///   emit(GetSuccessState<UserModel>(data: model));
 /// }
 ///
@@ -53,11 +52,11 @@ abstract class GetController extends BlocController<GetStates> {
     if (autoRun) run(args);
   }
 
-  /// A neccessary function to override when we extends this controller.
+  /// A function where we execute api request.
   ///
   ///```dart
   /// @override
-  /// Future<void> request() async {
+  /// Future<void> onRequest() async {
   ///   Response response = await Response.get('https://base.url/api/user',
   ///     onProgress: (double progress) {
   ///       emit(GetLoadingState<double>(data: progress));
@@ -68,13 +67,25 @@ abstract class GetController extends BlocController<GetStates> {
   ///   emit(GetSuccessState<UserModel>(data: model));
   /// }
   /// ```
-  Future<void> request(Map<String, dynamic> args);
+  Future<void> onRequest(Map<String, dynamic> args);
 
-  /// now, when we emitting the [GetStates] don't forget to define the object
+  /// A function that will be called when [onRequest] throws an error.
+  ///
+  /// ```dart
+  /// @override
+  /// Future<void> onError(dynamic e, StackTrace s) async {
+  ///   emit(GetErrorState(message: '$e'));
+  /// }
+  /// ```
+  Future<void> onError(dynamic e, StackTrace s) async {
+    emit(GetErrorState(message: '$e', data: s));
+  }
+
+  /// Now, when we emitting the [GetStates] don't forget to define the object
   /// type to emphasize the data that we're going to use in [ApiBloc].
   /// ```dart
   /// // in controller
-  /// Future<void> request() async {
+  /// Future<void> onRequest() async {
   ///   emit(GetSuccessState<UserModel>(data: model));
   /// }
   ///
@@ -94,12 +105,14 @@ abstract class GetController extends BlocController<GetStates> {
   void emit(GetStates<Object?> value) => super.emit(value);
 
   @override
-  Future<void> run([Map<String, dynamic> args = const {}]) async {
+  Future<void> run([
+    Map<String, dynamic> args = const {},
+  ]) async {
     emit(const GetLoadingState());
     try {
-      await request(args);
-    } catch (e) {
-      emit(GetErrorState(message: '$e'));
+      await onRequest(args);
+    } catch (e, s) {
+      await onError(e, s);
     }
   }
 
@@ -110,4 +123,55 @@ abstract class GetController extends BlocController<GetStates> {
 
   /// Wether the controller should trigger run on initialization or not.
   final bool autoRun;
+}
+
+class ApiFetcher extends GetController {
+  ApiFetcher({
+    required Future<void> onRequest,
+    Future<void> Function(dynamic e, StackTrace s)? onError,
+    super.autoRun = true,
+    bool autoDispose = false,
+  })  : _onRequest = onRequest,
+        _onError = onError,
+        _autoDispose = autoDispose;
+
+  final bool _autoDispose;
+  final Future<void> _onRequest;
+  final Future<void> Function(dynamic e, StackTrace s)? _onError;
+
+  @override
+  Future<void> onRequest(Map<String, dynamic> args) => _onRequest;
+
+  @override
+  Future<void> onError(e, StackTrace s) {
+    return _onError != null ? _onError!(e, s) : super.onError(e, s);
+  }
+
+  @override
+  bool get autoDispose => _autoDispose;
+}
+
+class SendFinalController extends SendController {
+  SendFinalController({
+    required Future<void> onRequest,
+    Future<void> Function(dynamic e, StackTrace s)? onError,
+    bool autoDispose = false,
+  })  : _onRequest = onRequest,
+        _onError = onError,
+        _autoDispose = autoDispose;
+
+  final bool _autoDispose;
+  final Future<void> _onRequest;
+  final Future<void> Function(dynamic e, StackTrace s)? _onError;
+
+  @override
+  Future<void> onRequest(Map<String, dynamic> args) => _onRequest;
+
+  @override
+  Future<void> onError(e, StackTrace s) {
+    return _onError != null ? _onError!(e, s) : super.onError(e, s);
+  }
+
+  @override
+  bool get autoDispose => _autoDispose;
 }